Initial Commit
This is a web scraper for the site: prnt.sc/ that generates a random
working url from this site, then pulls it and displays the image
(currently just the url in text).This is the very basic part of my
project. I intend on making it into an actual web page that displays
these random images.